use uni hai/weth pool price in header/footer and analytics graph
