+++ src/providers/AnalyticsProvider/index.tsx
@@ -1,7 +1,8 @@
-import { createContext, useContext } from 'react'
+import { createContext, useContext, useMemo } from 'react'
+import { formatEther } from 'ethers/lib/utils'
 
-import type { ReactChildren } from '~/types'
-import { Timeframe } from '~/utils'
+import type { ReactChildren, SummaryItemValue } from '~/types'
+import { Timeframe, formatSummaryValue, stringsExistAndAreEqual } from '~/utils'
 import { type HistoricalStatsReturn, useHistoricalStats } from './useHistoricalStats'
 import { DEFAULT_ANALYTICS_DATA, type GebAnalyticsData, useGebAnalytics } from './useGebAnalytics'
 import { type SystemData, useSystemData } from './useSystemData'
@@ -15,6 +16,7 @@ type AnalyticsContext = {
     haiPriceHistory: HistoricalStatsReturn
     redemptionRateHistory: HistoricalStatsReturn
     pools: PoolAnalytics
+    haiMarketPrice: SummaryItemValue
 }
 
 const defaultState: AnalyticsContext = {
@@ -40,6 +42,10 @@ const defaultState: AnalyticsContext = {
         loading: false,
         error: '',
     },
+    haiMarketPrice: {
+        raw: '',
+        formatted: '$--',
+    },
 }
 
 const AnalyticsContext = createContext<AnalyticsContext>(defaultState)
@@ -60,6 +66,22 @@ export function AnalyticsProvider({ children }: Props) {
 
     const pools = usePoolAnalytics()
 
+    const haiMarketPrice = useMemo(() => {
+        if (!pools.uniPrice) return data.marketPrice
+        const collateral = data.tokenAnalyticsData.find((data) =>
+            stringsExistAndAreEqual(data.tokenContract, pools.uniPrice?.token1)
+        )
+        if (!collateral) return data.marketPrice
+        const price = parseFloat(pools.uniPrice.token0Price) * parseFloat(formatEther(collateral.currentPrice))
+        return formatSummaryValue(price.toString(), {
+            minDecimals: 4,
+            maxDecimals: 4,
+            minSigFigs: 2,
+            maxSigFigs: 4,
+            style: 'currency',
+        })!
+    }, [pools.uniPrice, data.tokenAnalyticsData, data.marketPrice])
+
     return (
         <AnalyticsContext.Provider
             value={{
@@ -70,6 +92,7 @@ export function AnalyticsProvider({ children }: Props) {
                 haiPriceHistory,
                 redemptionRateHistory,
                 pools,
+                haiMarketPrice,
             }}
         >
             {children}
